data: to_guess now requires both axes to be set
Previously to_guess would return None for each unset axis, but now it
raises a DataErr, since the assumption is that there is valid data to
guess. There is an attempt to remain compatible with the existing
error messages, but there may be subtle changes (normally a change
in the DataErr message, but this is not guaranteed [*]).

[*] e.g. DataPHA("x", [1, 2, 3], None).to_guess() would have errored
    out as it tried to divide an array by None. It now raises a
    DataErr with a message that the dependent axis has not been set.