data: to_guess/fit/plot now requires both axes to be set
Previously to_guess would return None for each unset axis, but now it
raises a DataErr, since the assumption is that there is valid data to
guess. There is an attempt to remain compatible with the existing
error messages, but there may be subtle changes (normally a change
in the DataErr message, but this is not guaranteed [*]).

[*] e.g. DataPHA("x", [1, 2, 3], None).to_guess() would have errored
    out as it tried to divide an array by None. It now raises a
    DataErr with a message that the dependent axis has not been set.

Similar changes are made to to_fit and to_plot. There is also the
removal of a to_fit method in DataPHA since it didn't add anything
compared to its superclass.

The Fit.fit and .est_errors methods have been reworked to handle the
new behaviour but retain the old behaviour - e.g. to throw a DataErr
or FitErr. This has meant we can remove the internal _check_length
routine, as it's no-longer needed.