Apply cache to call 2D model, but set size to 0 by default
This commit applies the cache to all models, even those where it does not make sense (e.g. a simple constant). However, this way, we get a consistent interface.
Then, we switch the default of all cache sizes for 2D models to 0, so that the user needs to explicitly activate them if needed. In 2D arrays are often bigger, so there might be memory implications of using the cache and thus we don't want it to activated unless asked for.