TEST: add a basic test of plot_rmf
There are a number of tests of the RMF plotting, but they all check
get_rmf_plot. Add an actual test of plot_rmf. This is for the
matplotlib backend, so the actual plot data can be checked, but only
minimal checks are made, as the assumption is that if these tests pass
then the other tests - e.g. of get_rmf_plot - cover the other
cases (e.g. the actual data plotted).