sim: add clip support to normal/uniform/t_sample
This follows the approach of #866 to default to applying the hard
limits to the parameter samples *before* they are used to calculate
the statistic values (technically it can be done afterwards since
the values should get clipped to the hard limits anyway, but this
is a lot cleaner, and allows for the clipping to the soft limits).

Unfortunately the way clipping was introduced in #866 does not
work here, since we want the clipping done before the statistic
values are calculated. This leads to subtly-different behaviour
(which could perhaps be reworked once all the code supports
clipping).

Another difference to #866 is that we do not return any information
about whether a row was clipped or not. I did not think it worth
passing back this extra information as we currently just have a
single ndarray. If the return type were more structured then it
might be added, but is it worth it?

add clip to normal_sample/t_sample