TESTS: add basic sample_flux tests for #2337
This is to check the behaviour with a linked parameter, but we take
the time to add some extra checks of this condition. The assumption
is that most features of the sample flux code are already tested.

An assert condition has added a label so that it's obvious when
it is hit (as happens in one of these tests).