Add an interface to `scipy.optimize`
Scipy is a widely supported pacakge in the sciencific Python ecosystem and
contains a number of optimiation methods not available in Sherpa itself.
Since each method has advantages and drawbacks, it is useful to offer
users are wider range of opmization methods for choose from for their
particular problem.

The implementation here adds an interface for Scipy and genrates
Opitimizer classes, if scipy is availble in the environment.
That way, there is no change in behavior for users that do not have
scipy installed.

Compared to the built-in optimizers, there is limited testing here,
because the optimization methods themselves are already  tested by
scipy. Here, we only need to ensure that the interface works.