Docs: improve parallel documentation
The main aim was to mention that we do not expect Sherpa to benefit
from the free-threading work, but I di not want to add warnings
about this throughout the whole code base as it is not clear where
best to add it.

As part of this, add documentation about the configuration file.