CI: replace the arch workflow with the new aarch platform
The arch workflow, which is run weekly, is a test on Linux/aarch
platforms. However, as it is not currently a supported platform, we so
not run the tests on every commit, and because these tests take quite
a long time to run (as it uses emulation).

Now that GitHub has released a direct linux aarch platform, replace
the emulated runs with a direct one. This does mean we lose checking
one platform (ppc64) but the reduction in test time is worth it.

The conda setup code has been updated to work with this platform, even
though it is not used in this commit (since the changes have been made
to the pip workflow).

As part of this, the arch CI run now includes DS9 in its testing.