Rewrite example to simulate X-ray data
In Sherpa 4.16.x the interface to `fake_pha` has changed considerably, with several options now out of use and instead it's up to the user to encode some of that information (for example the response) into the model that is passed in.
The docstring of `fake_pha` was updated at the time, but not the separate narrative pages in the docs. Thus, all the doctests in there fail (which we missed because we dn't run them by default - and of course that's exactly why we should).
While rewriting anyway, the images in the docs are not automatically generated. That serves two purposes: First, they need to be remade anyway to be in line with the updated text and, second, that means at least the doc build (e.g. readthedocs on CI) will fail if further updates to `fake_pha` require changes to the examples before we activate the doctests by default.
To make this work, though, a small change in configuration in needed: To make sure RTD has the datafiles to run the examples, we add the submodule sherpa-test-data to the readthedocs built, get the datafile for the docs built itself, and add a line of text for users who wonder where the datafile comes from.

closes #2122
closes #2043