xspec: remove eval_xspec_with_fixed_norm decorator
The eval_xspec_with_fixed_norm decorator was added in #2268 to take
advantage of possible cache gains (since the normalization parameter
of XSPEC models is a linear correction it means it can be cached
without the norm parameter, potentially speeding up cases where only
the normalization changes). However, we still needed to send a
fake normalization value through to the code (the value being 1).

With the new approach the "fake normalization" value is not needed,
which ends up removing the need for the decorator since the logic can
just be hard-coded in the calc method of the additive model class.