TESTS: regression test for plot_fit_xxx label handling
If plot_fit_resid (and presumably the other variants) is called, check
what happens to the plot objects for data, model, and residual. Are
the labels adjusted to match those used in the actual plot_fit_resid
call, or left as is?

This is a regression test as there are reasons to behave both
ways (e.g. match the plot_fit_resid call or match the individual
plot_xxx calls).