activate doctests by default
This requires making sure that doctestplus is installed even in a call to
sherpa_test, because otherwise the "doctest-rst" option would raise an error.
It also requires small changes to the get_datadir mechanism and the import
of sherpatest to ensure the doctests that rely on the data are skipped when
it's not present.