UI: allow plot calls to use multiple identifiers
Allow

    plot("fit", [1, 2])

to match

    plot_fit([1, 2])

Unfortunately the way the code works makes the plot code hard to write
while remaining generic. In particular:

- we do not support multiple ids for all plots, so some are
  not expected to work

- we can not guarantee the order of the arguments (e.g. the
  first argument is not guaranteed to be an identifier)

They are symptoms of the same issue - e.g. how to handle the
"model_components" plot option - which is left for a different time to
think about, since leaving as is maybe the most sensible option.

The way that multiple arguments to a keyword - such as

    plot("data", [1, 2], "model", [1, 2], color=["r", "b"])

is potentially confusing.

WIP