UI: allow per-plot keyword arguments for plot
Per-plot keyword arguments can be sent in via a dict, so

  plot("data", {"color": "green"}, "model", { alpha: 0.5 })

This turns out to be easy to accomodate, but it would be nice to
have a better way than an "isinstance(v, dict)" check. However,
we have to be careful because

  plot("model_component", 2, mdl)

is valid and models now behave somewhat dict-like, so it is safest
(and easiest) just to use the isinstance check.