TESTS: support unexpected test failures on CI
The WARNING messages that get caught by caplog has not been clear
to me, and the tests behave differently on different systems.

Change the tests to allow for these extra messages if they are
present. Since the messages are not the main reason for the test
they are not all checked for (i.e. we do not necessarily check
what these extra messages are).