Reorganize class hierarchy for Dataclasses and EvaluationSpaces
This PR is the fist step of a reorganization that aims to bring more consistency. Data1D and Data2D and their integrated counterparts are the most used dataclasses in Sherpa. Yet, their inheritance is somewhat murky and their behavior somewhat inconsistent. This stems from the hierarchy in which they are developed: `Data2DInt` derives from `Data2D`, but `Data1DInt` is nowhere in the ancestors. Instead the "integrated" bits are implemented separately. Similarly, the `DataND` class actually has no counterpart for integrated datasets.

In this PR, "integrated" simply becomes an option for `DataND` and then all classes specialized on one particular dimensionality inherit the ability to deal with integrated data or not; in the case of the existing data classes `Data1D` and `Data1DInt` that means that they just call `DataND` with integrated set to True or False, as needed.

This commit aims to be almost entirely backwards compatible (two minor changes are explained below), which means that some methods are implemented more than once to keep the previous inconsistent behavior: For example, in 1D "overlaps" is defined as "the space overlaps somewhere", while in 2D " overlaps means "The Evaluationspaces have the exact same endpoint in every dimension".

In later commits, I will look at breaking backwards compatibility to achieve a more consistent behavior.

Changes in this commit (see changed tests):
- A field "integrated" is now printed out in the string reprensentation. That is useful, because the same class can now represent an integrated or a non-integrated dataset. For the legacy classes this information is also in the class name, but it doesn't hurt to also print it out.
- The format of an empty grid returned from a 2D class has changed from a tuple of arrays with entry `None` to a tuple of `None` values. This could be fixed with yet another special treatment of None values in the axes, but the change makes is more consistent and seems to have minimal impact, thus I'm including it in this commit.