remove `for_model` method
The `for_model` method was unique to Data1D/Data1DInt` and not available in any other dataclass. It was also useless: Careful examination shows that it will always return `self` independent of the input parameters. It look like there are originally intended to be another branch of a an if statement that would return the evaluations space attached to a model, if such space existed, but without that branch, the function is just returning self.

It was also called from within `get_evaluation_indep`. Removing the call to `for_model` means that now `get_evaluation_indep` and the simple `get_indep` are identical except for the calling signature, so `get_evaluation_indep` can also be removed and replaced with `get_indep`. This change brings more consistency to the Dataclasses for different dimensions.