With the removal of `get_evaluation_indep` the keyword `use_evaluation_sapce` in `get_x`, get_y` is now useless, so it has been removed.
We could leave it in (at least in the function definition) for backwards compatibility, thus I made this a separate commit so that it would be easier to modify if we wanted to do so.

(Similarly, the model does not have to be passed into `get_x` any more, but that is used in more places, so removing it would be more controversial and I'm leaving that for a separate commit).