TESTS: fix configuration error
The pattern for recognizing numbers used by check_str turned out
to be invalid (missing a "," to separate terms) but until now we
never ended up hitting the problem. There was also a problem since
the "not-followed-by" check did not include the decimal point,
causing problems for certain cases).

Also fix the code to avoid using the match symbol as this is now
valid Python.