Change class hirachy of dataspaces and evaluationspaces
This is a huge individual commit, because changing the class hirachy oa the bottom of the causes ripples that
require changes all the way through. Yet, by number of lines, most changes are not actually new code, but code moved around,
for example from `Data1D` into the new `BaseData1D`.
The gaol is to have `EvaluationSpaceND` be the base of all evaluation spaces.
`EvaluationSpace1D` and `EvaluationSpace2D` are merely specializations of that with a slightly different calling structure
(The independend axes are passed individually to the constructor instead of as a tuple).
An evaluation space can be flat (i.e. all axes have the same length such as `x0=[1,1,1,2,2,2]` and `x1=[1,2,3,1,2,3]`)
or not (`x0=[1,2]` and `x1=[1,2,3]`). The former case allows evaluation on a non-rectangular grid.

Data spaces derive from Evaluation spaces by adding a filter as a mix-in class. This design allows us to reuse
the entire inheritance tree and basically build a parallel tree by just mixing in the mix-in class in every case.
It also means that we can reproduce our legacy DataXD and DataXDInt (with X=1,2) classes with very little code.

Because this commit is already so big, we attempt to keep API changes to a minimum, we can be seen as a very small
number of changes in the tests. Further changes will be made in follow-up commits to simplify and streamline a few
features, to make use of this new structure, but here we attempt to preserve the API as much as possible.
It turns out that there are a few cases where it would be very cumbersome to keep the exact behavior and we intend
to change it in the follow-up commit anyway, those are lsited below:

This also changes the way that the size of a dataspace is validated. It now refers to the size of the underlying Evaluation space if set already,
and if not the size of the the independent axis instead of inpdenently trackign a size.

That means that it is now an error to set a "related" field (e.g. the dependent axis or the error) without setting the independent axis first.

However, there is a difference in setting the independent axis to `None` or `[]`. The behavior is the same, but the error messages are different.