Add a  convenience method to set the IO backend and a context manager to set the IO backend temporarily.
This is implemented analogously to the plotting backend, but differs in the details because the IO backend is a module, not a class.

This functionality is probably mostly useful for developers; users generally don't change the backend once it works since they all function identically (unless there is a bug). This is in contrast to the plotting backends, where a user might hae good reasons to change between interactive and statics plots within one session. Thus, the docs for the new feature are more muted (it's just in the module docstring and the docstrings for the functions, not in the narrative docs).

Fixes #2101