TEST: upate test to run with crates backend
PR #2352 removed support for XSPEC table models from the
load_table_model command, and we updated the test to check this.
However, it turns out that the behaviour depends on the I/O backend,
and as we do not have a crates CI run we did not notice this. I have
now fixed the test.