Allow "dummy" as an alias for BasicBackend in plotting
I tested this locally and I don't think it needs and automated test.
However, the question is if we want to make this change at all. It's motivated by #2167, but we've now had a full release cycle without "dummy" as a valid plot backend value, so it is still worth it to add the alias back in now?

Fixes #2167