Pass backend-specific options through to pylab and bokeh
Sherpa plotting commands should be passing through keywords that are not part of the sherpa-specific interface to the plotting library that is used as backend underneath.
This was one of te main design goals of the plotting work done for 4.17, and there is even a test for it, but as it turns out that test happens to use the wrong keyword (one that is part of the sherpa-specific set) so we never caught that it was, inf fact, NOT passing through stuff.

Two related small changes are included here:
- remove a print statement that was left over from some debugging
- There is an ambiguity on how to match general (sherpa API) keywords to backend keywords when it's not a 1:1 ratio. For example, if alpha is set, should it apply to the markers, the lines, the hatches or all of them? Hatches are invisible when they are always exactly identical in color, alpha etc to the filling color of a symbol, so we are no longer setting the hatch properties based on the general input and leave it to the user to pass through hatch specific keywords.

fixes #2167

This should have been part if 4.17, so it would be great to get it into 4.18 at least.