Remove duplication of the `load_table_model` method between `sherpa.ui` and `sherpa.astro.ui`.
To implement this I align the signature of the `load_table_model` in `sherpa.ui` and `sherpa.astro.ui`.
This is technically an API breaking change in case someone relied on the ordering of arguments and passed in the interpolation method as third argument instead of a keyword argument.

I think that's unlikely to happen in practice (most people won't pass that in at all and just use the default), but it's separated into it's own commit in case we want to discuss or reverse that change.

Matching the API does not only make the linters happy, it also simplifies documentation as we could now point to (or generate form in some way) the docs of this method from the docs of the superclass  method.

In fact, I'm removing keywords that are no longer used, because they are confusing. Agian, I don't think any of them are used in practice, but to keep backwards compatibility, we could just keep those unused keywords in the signature and raise a DeprecationWarning if they are set.

closes #1648