UI: calculate covariance within get_draws; fix #556
We assume that covar() is not an expensive call to make, so we
now *always* calculate the covariance matrix within get_draws, rather
than rely on the user having called covar first.

This should make things a bit easier for the user, as well as fixing
problems like #556 - where it was possible to get the wrong answer by
not calling covar first.

There is the question of whether the user should see this covariance
output, but I decided against it.

We also do not store the results of the "covar" call, since this could
confuse existing code (although if it did this would suggest that the
code was actually subtly wrong).

A different option would be to identify when the covar results are
not valid and only call covariance in this case, but that turns
out to be more awkward than it seems, and it is just easier to
always re-calculate the matrix.