UI: save_all improvements for PHA2 files. Fix #1882
We now store information when loading a PHA2 file (identified as
the only load command that can load in multiple data files from a
single call). This information is then used to avoid writing out
load_pha/bkg calls that end up being wrong.

I believe that this technique - that is, store the load data call
information - should be expanded for all load calls. However, this is
a large change that needs work, since there are cases where the new
code can fail (but it is expected that users who are doing this are
capable of writing out their own Python code to serialize the data).