UTILS: internal changes to eqwidth code
There should be no functional change to the code.

A few style changes

- change Optional[x] to x | None
- avoid setting variables to None when they are about to be
  over-written
- set num once
- move variable definitions closer to their usage (num and eqw)
- tweak a loop that was calculating an unused variable
- add some comments on things we could look at in the future,
  mainly about the use of non-PHA data

Minimal documentation has been added. Unfortunately there is some
confusion over what the code is doing (and so what the variables
are). The current documentation is based on that I wrote for the
UI version of eqwidth, but it does look like this equivalent
width is calculating the negative of the version described
on the wikipedia page.

WIP