UTILS: internal change to pha serialization
Track files that are automatically included with PHA files. This
is to try to simplify the output of save_all but at present doesn't
actually do anything (since PHA2 files do not auto-load responses).

As part of this, add support to track data deletions and data changes
(via set_data and copy_data).