UI: improve background handling for serialization
Store the load_bkg data as is done for PHA data with load_data/pha.
This is slightly different, since we can have multiple entries
(bkg_id) for a given idval, necessitating a slightly-different
structure.

The stored background data can be used to slightly simplify the
PHA serialization (rather than assert "we haven't loaded this
automatically" we can now assert "we did load this explicitly").

Finishes the fix for #2383.