TESTS: check pha serialization with the use_errors flag
The code currently ignores the use_errors flag, so add a test that
explicitly checks this behavious (and so is marked xfail) and one that
is a regression test (so the serialization output will need to be
updated once this bug is fixed). See issue #2383.

We also add a check of copy_data() working on PHA data - that is,
do we track enough information to allow

    load_pha(1, infile)
    copy_data(1, 2)
    delete_data(1)

to be represented? As part of this, add a use_errors=True to the load
call to check this keyword gets passed through with the copy.