TESTS: check serialization of manually-created responses
At the moment save_all does not handle this case well, so add a test
that checks it fails as expected, so any change will be noted by the
tests.

The test needs to be marked as @requires_fits since the incorrect
serialization ends up adding load_arf/rmf calls to the file, so
we need a FITS I/O backend to run the test.