UI: add auto_load argument to save_all
Recent changes to save_all mean that the default behavior is to
no-longer include load_arf/rmf/bkg lines for "automatically loaded"
ancillary files (as Sherpa will load them for you).

However, there is some use in having the explicit version, as it makes
it obvious what files are needed by the script. So, add a flag to
save_all - auto_load - which, when set to False, will re-create the
old output (hence include the load_xxx calls).

Fortunately, due to the way that the serialization picks up what files
are automatically loaded, this turns out to be easy to implement.

WIP