STATS: remove CStatNumpy class
There is no benefit at this time to keeping the CStatNumpy class,
as - other than providing an example of how to implement the
statistic calculation directly in the class, it doesn't do anything
different to CStat. As we have CStatNegativePenalty which also
implements the statistic calculation in Python there is no reason
to keep it around.