DOCS: add information about CStatNegativePenalty to RTD
This takes some of the original information from the
CStatNegativePenalty docstring and moves it into the RTD
documentation.

This uses the example from https://github.com/sherpa/sherpa/issues/836
since the simlpified case that @hamogu found is a bit hard to
use as an example (because the starting point needs to be
negative, which is hard to say why you start the fit there).

As part of this, add a new doctest fixture in doctest/conftest.py
to add the location of the documentation code to the namespace. This
is needed to allow the fit/poisson.rst code to be able to find the
correct files for the tests.