UI: improve documentation for eqwidth
We have previously linked to the wikipedia artical on equivalent
width, but this turns out to to be the negative of what the code
calculates. It turns out - see ussue #2381 - that we actually
implement the approach from George & Fabian (1991). If we have

  c = continuuum
  l = line

then we have sum(l * bin_width / c). However, the code is written
"more generally", where we send in two models:

  c_m - the continuum model
  x_m - the combined model (so continuum + line)

and in this case we can assume that

  c_m = c
  x_m = c + l

and so

  (x_m - c_m) / c_m ==
  (c + l - c) / c   ==
  l / c

which is what George & Fabian (1991) have (modulo the bin-width
as this is really the term inside the integral from the wikipedia
page, but George & Fabian only talk about a single bin).