TESTS: fix xxx_sample tests for change in default clip value
Set clip="hard" since this is no-longer the default. These tests
are not really checking the clip behaviour, so it is easier to
make this change rather than re-calculate all the "expected"
values (most of these are regression tests so we are just looking
for changes in the returned values, not checking the values are
the expected values, since most of this depends on the RNG used).