XSPEC: add initial support for XSPEC 12.15.1
Allow Sherpa to build against XSPEC 12.15.1.

This provides low-level access to the new models but it does not
add the new models, which are:

  XSbfekblor     XSzbfekblor
  XSbfeklor      XSzbfeklor
  XSfekblor      XSzfekblor
  XSrsapec
  XSrsgaussian
  XSrsrnei
  XSrsvapec
  XSrsvrnei
  XSrsvvapec
  XSrsvvrnei

XSPEC 12.15.1 appears to have added support for specifying the
ATOMDB/NEI/SPEX versions to "latest" (case insensitive).

The implementation appears to differ depending on whether you
are using XSPEC or the XSPEC model library (this is presumably a
bug).

Of course, for Sherpa - which supports multiple XSPEC versions -
we have to decide

- when to worry about "latest" settings,
- and what to use for the defaults.

Since users can end up with "latest" values even when building
Sherpa with "old" XSPEC versions - as the value can be read from
~/.xspec/Xspec.init - we decide to always try to convert "latest",
and use the HEADAS 6.36/XSPEC 12.15.1 values for atomdb, nei, and
spex as the default (spex only set when XSPEC >= 12.15.0).

It looks like the APECROOT/NEIAPECROOT/SPEXROOT string-database
settings now interact with the atomdbVersion/neiVersion/spexVersion
settings (i.e. changing one changes the other). This makes our
life easier as there is less need to try and synchronize settings,
at least if you have XSPEC 12.15.1.

There are hidden changes to how the "string database" as handled by
set_xsxset/get_xsxset behave: 12.15.1 now appears to always add
several keywords (APECROOT, NEIAPECROOT, and SPEXROOT) which ends up
causing problems for our save_all/serialization tests. So the tests
have been re-written to just check that the values are set, but
not care about the actual value (via the "# doctest: +ELLIPSIS" option
added here).

As part of this work, update the installation documentation for
configuring XSPEC:

- change the labeling of the per-version guides from numeric to
  unlabeled (to simplify updates when adding new or removing old
  versions)

- add a note that if XSPEC was installed via conda then the
  settings are (likely) simpler.