XSPEC: add initial support for XSPEC 12.15.1
Allow Sherpa to build against XSPEC 12.15.1.

This provides low-level access to the new models but it does not
add the new models, which are:

  XSbfekblor     XSzbfekblor
  XSbfeklor      XSzbfeklor
  XSfekblor      XSzfekblor
  XSrsapec
  XSrsgaussian
  XSrsrnei
  XSrsvapec
  XSrsvrnei
  XSrsvvapec
  XSrsvvrnei

There are hidden changes to how the "string database" as handled by
set_xsxset/get_xsxset behave: 12.15.1 now appears to always add
several keywords (APECROOT, NEIAPECROOT, and SPEXROOT) which ends up
causing problems for our save_all/serialization tests. We attempt to
write this in a way that will work with newer XSPEC libraries, but
it's hard to know for sure.

As part of this work, update the installation documentation for
configuring XSPEC:

- change the labelling of the per-version guides from numeric to
  unlabelled (to simplify updates when adding new or removing old
  versions)

- add a note that if XSPEC was installed via conda then the
  settings are (likely) simpler.