SCRIPTS: improve XSPEC model creation code
Model parameters are now labelled as "Attributes" rather than
"Parameters" in the docstring as

a) this is true
b) it avoids sphinx errors for certain cases (e.g. "vv"-style
   models where we want to label all the parameters with the same
   documentation string)

The add_xspec_model script will now error out if the model is
already known about, but a command-line flag can be used to turn
off this check).