XSPEC: initial support for "latest" atomdb/nei/spex settings
XSPEC 12.15.1 appears to have added support for specifying the
ATOMDB/NEI/SPEX versions to "latest" (case insensitive).

The implementation appears to differ depending on whether you
are using XSPEC or the XSPEC model library (this is presumably a
bug).

Of course, for Sherpa - which supports multiple XSPEC versions -
we have to decide

- when to worry about "latest" settings,
- and what to use for the defaults.

Snice users can end up with "latest" values even when building
Sherpa with "old" XSPEC versions - as the value can be read from
~/.xspec/Xspec.init - we decide to always try to convert "latest",
and use the HEADAS 6.25/XSPEC 12.15.0 values for atomdb, nei, and
spex as the default (spex only set when XSPEC >= 12.15.0).

There is uncertainty frmo XSPEC over the default spex value - should
it be "3.08" or "3.0.8" - and for now we go with teh former (since
using the latter causes down-stream problems).

It looks like the APECROOT/NEIAPECROOT/SPEXROOT string-database
settings now interact with the atomdbVersion/neiVersion/spexVersion
settings (i.e. changing one changes the other). This makes our
life easier as there is less need to try and synchronize settings.
At least if you have XSPEC 12.15.1.