XSPEC: query and set the database-library versions
There are multiple ways of accessing the ATOMDB/apec, NEI, and SPEX
database versions. As of XSPEC 12.15.1 it appears they are correctly
handeled (e.g. so setting atomdbVersion will also change the
"ATOMDB_VERSION" model setting).

Allow the databse versions to be queried by get_xsversion and set by
the new set_xsversion call. Rather than overload set_xsversion we
could have added a new call, but it doesn't seem worth it.

Note that support is complicated by the fact that some of these
libraries are only supported with XSPEC >= 12.15.0 ("spex").

This code is based on PR #2216 (which itself was likely originally
created in an older PR).