XSPEC: correctly label model parameters as Attributes
This is a change to the docstrings for models, pointing out that
the model parameters are attributes, rather than parameters used
when creating a model object.

The mean reason for doing this is that it improves the Sphinx output
for these models, and avoids some issues for "vv" models where we want
to use the same explanation for many attributes (e.g. He to Zn).

The string "Parameters" was incorrectly suggested by the
add_xspec_model.py script.