New Table models to split functionality of existing TableModel
TableModel really has two separate applications: It can be an x/y table that is interpolated in x as needed, or it can just be a table of y values assuming that the number of values exactly matches the number of data values.

This is conceptually different and can be confusing to users (and also in the code). This commit introduces two new classes to split the two different functionalities into separate classes (with a common base class).