regrid: force an empty-axis error for .start/.end properties
The existing code allows axes to often be unset, but then this can
lead to errors, such as trying to access the first element of the
array which doesn't exist.

To guard against this, add some explicit error calls for the start and
end properties of Axis objects. This is not that useful until we add
typing rules to this module, but does indicate intent.