sim: add type statements
Add type statements to the top level of the sherpa.sim module.

There is a lack of clarity of whether we are sending in a class or
an instance of a class - e.g. the MCMC set_sampler routine takes
a class, but in other cases we send in the instance - which makes
this a bit tricky to do.

There is also some attempt to "modern-ize" the class structure
by using ABC as the base class (but it is only a minimal change
and should not affect users).