image/DS9: start simplifying DS9 interface
Remove the showFITSFile method (and the commented-out showBinFile) as
they are not used by Sherpa. This means that we can also remove the
_expandPath routine. As part of the cleanup remove the unused
_addToPATH routune (which, on review, contained logic errors anyway).

The intro text has been cleaned up to remove the list of changes, as
this has not been kept up-to-date and is better found using git
nowadays anyway.