image/DS9: simplify module initialization; drop setup routine
The module checks if both ds9 and xpaget executables can be found and
will error out if they can not be. It attempts to support the ability
to re-check this and "turn back on" the module: the intention is to
allow a user to install the ds9 and/or xpa tools and then the code can
work.

However,

- this is not used by Sherpa; as soon as the module fails it remains
  "failed" (that is, there's no "reload" functionality unlike the
  I/O and plotting backends),

- it complicates the code (e.g. the use of global variables which may
  be undefined causes pyright issues when checking types),

- and it is not exactly obvious how a user is meant to use this
  feature **without manually editing the DS9.py file** since importing
  the module is always going to fail, so the user can never access the
  setup call to run with doRaise=False.

So, remove the complexity and just deal with whether ds9 and xpaget
can be found at module import time. As part of this remove the setup
routine.