image/DS9: drop the doRaise argument
Sherpa never sets doRaise to anything but True, so fix this behaviour.
It does mean that a user can not start Sherpa, find they do not have
DS9, install it, then reload / rerun the code to pick it up
without restarting the Python process (or a more-involved reload/import).
However, this behaviour is not expected with Sherpa so it is better to
not have vestigal code left around.