image: fix minor issue with OBJECT name behaviour
The name field of DataImage will get over-written if the data object
has a header with an "OBJECT" keyword. However, this was never
"reset", meaning that it would persist for any DataImage display until
a new data object with an "OBJECT" keyword is found. The prepare_image
call now explicitly resets the name field in this case.