utils: simplify bool_cast and add other boolean type statements
The bool_cast routine has likely long-passed it's purpose - as it was
likely added to support users used to the pre-Python Sherpa as well as
those using Sherpa via the S-Lang bridge. However, it is a bit early
to remove, but we can simplify its implementation as it turns out the
only place that we ever exercise the "convert a sequence of values to
a ndarray of bools" is in a single test. So we can just leave it as
working on scalar values, which greatly reduces the number of type
complaints when used in other parts of the system.

To address review comments, the is_iterable and is_iterable_not_str
routines have been removed from sherpa/utils/__init__.py and either
inlined directly or moved to the module that uses it (for
sherpa/fit.py and is_iterable).

With pyright 1.1.408, the number of typing errors decreases due
to these changes (mainly the bool_cast simplification):

           ui/utils   astro/ui/utils   data   astro/data
  before:    117          160           88       162
  after:     106          131           87       154